+++ src/client/java/io/github/aaronateataco/petsandpals/enums/PetList.java
@@ -70,6 +70,7 @@ public enum PetList implements NameableEnum {
     stray,
     wither_skeleton,
     tadpole,
+    tropical_fish,
     turtle,
     vex,
     villager,